server/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cryptography==3.4.7
Flask==2.0.0
idna==2.10
IPy==1.1
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.0.1
MarkupSafe==2.0.0
numpy==1.20.3
panda==0.3.1
pandas==1.2.4
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.3
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
threadpoolctl==2.1.0
urllib3==1.26.4
Werkzeug==2.0.0
whois==0.9.13
